refactor(flow): split the contract-snapshot commands out of flow-cli
flow-cli.ts crossed the 500-line module threshold when the landing render
landed (the self-guardrail caught it). The refresh/publish/landing commands
are a cohesive unit — they move to flow-contract-cli.ts; the map/trace/
console views stay. Shared helpers (gatherModel, splitPaths, emitJson) are
exported from flow-cli rather than duplicated.
